Update roadmap in the README
We've done some of these things, and some of them we decided not to do.

Also, this commit includes a 'Sem-Ver' commit footer. This tells pbr
that when it is generating version numbers it should consider that this
is development towards 3.0.0 rather than development towards 2.5.3
(which is a release we do not intend to make) That will cause:

  Zuul version: 2.5.3.dev1212

at the bottom of the status page to change to:

  Zuul version: 3.0.0.dev1212

Change-Id: Ic49d127d8624a4f56c5adf4c9fb67621b0ac6b93
Sem-Ver: api-break